brining in some sense and sensibility...
