+++ pystorms/__init__.py
@@ -1,4 +1,7 @@
 from pystorms.scenarios import *
 from pystorms.utilities import *
 from pystorms.environment import *
+from pystorms.networks import *
+from pystorms.config import *
+from pystorms.binaries import *
 __version__ = "0.1.2"
