+++ pystorms/scenarios/__init__.py
@@ -1,0 +1,4 @@
+from .scenario import scenario
+from .epsilon import epsilon
+from .gamma import gamma
+from .theta import theta
