+++ pystorms/scenarios/theta.py
@@ -1,0 +1,75 @@
+from pystorms.environment import environment
+from pystorms.networks import load_network
+from pystorms.config import load_config
+from pystorms.scenarios import scenario
+from pystorms.utilities import threshold
+import yaml
+
+
+class theta(scenario):
+    r"""Theta Scenario
+
+    Separated stormwater network driven by a idealized event.
+
+    Parameters
+    ----------
+    config : yaml configuration file
+        physical attributes of the network.
+
+    Methods
+    ----------
+    step:
+
+    Notes
+    -----
+    Performance is measured as the deviation from the threshold.
+
+    """
+
+    def __init__(self):
+        # Network configuration
+        self.config = yaml.load(open(load_config("theta"), "r"), yaml.FullLoader)
+        self.config["swmm_input"] = load_network(self.config["swmm_input"])
+
+        self.threshold = 0.5
+
+        # Create the environment based on the physical parameters
+        self.env = environment(self.config, ctrl=True)
+
+        # Create an object for storing the data points
+        self.data_log = {"performance_measure": [], "flow": {}, "flooding": {}}
+
+        # Data logger for storing _performance data
+        for ID, attribute in self.config["performance_targets"]:
+            self.data_log[attribute][ID] = []
+
+    def step(self, actions, log=True):
+        # Implement the actions and take a step forward
+        _, done = self.env.step(actions)
+
+        # Log the flows in the networks
+        if log:
+            self._logger()
+
+        # Estimate the performance
+        __performance = 0.0
+
+        for ID, attribute in self.config["performance_targets"]:
+            if attribute == "flooding":
+                __flood = self.env.methods[attribute](ID)
+                if __flood > 0.0:
+                    __performance += 10 ** 6
+            if attribute == "flow":
+                __flow = self.env.methods[attribute](ID)
+                __performance = threshold(
+                    value=__flow, target=self.threshold, scaling=10.0
+                )
+
+        # Record the _performance
+        self.data_log["performance_measure"].append(__performance)
+
+        # Terminate the simulation
+        if done:
+            self.env._terminate()
+
+        return done
