+++ pystorms/utilities.py
@@ -1,79 +1,4 @@
 import numpy as np
-import os
-import platform
-
-# Local Path
-HERE = os.path.abspath(os.path.dirname(__file__))
-
-
-def load_network(name):
-    r""" returns the path to the desired network.
-
-    Parameters
-    ----------
-    name : str
-        name of the network. *alpha*, *beta*, *gamma*, *epsilon* and *delta* are valid
-        keywords
-
-    Returns
-    -------
-    path : str
-        path to the network
-    """
-    if name == "alpha":
-        path = os.path.join(HERE, "networks/alpha.inp")
-    elif name == "beta":
-        path = os.path.join(HERE, "networks/beta.inp")
-    elif name == "delta":
-        path = os.path.join(HERE, "networks/delta.inp")
-    elif name == "epsilon":
-        path = os.path.join(HERE, "networks/epsilon.inp")
-    elif name == "gamma":
-        path = os.path.join(HERE, "networks/gamma.inp")
-    elif name == "theta":
-        path = os.path.join(HERE, "networks/theta.inp")
-    else:
-        raise ValueError("Undefined Network, please refer to the documentation")
-    return path
-
-
-def load_binary(name):
-    r""" returns specified swmm binary based on the operating system.
-
-    Parameters
-    ----------
-    name : str
-        name of the binary. refer to the documentation for the key words
-
-    Returns
-    -------
-    path : str
-        path to the binary
-    """
-    # Find the path to the library
-    op_system = platform.system()
-    if op_system == "Linux":
-        path = "/linux/libswmm5.so"
-    elif op_system == "Windows":
-        path = "/windows/swmm5.dll"
-    elif op_system == "Darwin":
-        path = "/macos/libswmm5.so"
-    else:
-        raise ValueError(
-            "Operating system not identified, please check the binary path or overwrite it. \
-                    Refer to the documentation for additional details."
-        )
-
-    # Choose the binary
-    if name == "pollutant_support":
-        path = "binaries/" + name + path
-        path = os.path.join(HERE, path)
-    else:
-        raise ValueError(
-            "binaries not found. Refer to the documentation for the list avaiable binaries"
-        )
-
-    return path
 
 
 def append_rainfall(
